+++ src/App.jsx
@@ -1,43 +1,16 @@
-import { useState } from 'react';
-import './App.css';
-import './pages/Login.css';
-import Login from './pages/Login.jsx';
-import Registration from './pages/Registration.jsx';
-import Messages from './Messages';
-import Gallery from './Gallery';
+import { Routes, Route } from "react-router-dom";
+import LoginPage from "./pages/Login";
+import HomePage  from "./pages/Home";
+import Registration from "./pages/Registration";
 
-const App = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-
+function App() {
   return (
-    <div className='main-layout'>
-
-      <div className={`main-messages ${isOpen ? 'open' : ''}`}>
-        <div className='message-header'> 
-          Messages 
-        </div>
-        <Messages />
-      </div>
-
-      <div className={`hamburger ${isOpen ? 'open' : ''}`} onClick={toggleMenu}>
-        <div className="burger"></div>
-        <div className="burger"></div>
-        <div className="burger"></div>
-      </div>
-
-      <div className='line'></div>
-      
-      <Gallery />
-      
-      <div>
-        <Registration />
-      </div>
-    </div>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+      <Route path="/login" element={<LoginPage />} />
+      <Route path="/registration" element={<Registration />} />
+    </Routes>
   );
-};
+}
 
 export default App;
