+++ src/Groups.css
@@ -1,0 +1,3 @@
+body {
+    background-color: black;
+}
