+++ src/pages/Login.css
@@ -1,10 +1,12 @@
+body{
+    margin: 0;
+}
+
 .input-container{
     display: flex;
     justify-content: center;
     align-items: center;
-    height: 100vh; /* Full height of the viewport */
-    width: 100vh;
-    background-color: #0d1538; /* Optional background color */
+    background-color: #0d1538;
 }
 
 .login-form {
@@ -13,7 +15,7 @@
     width: 300px; /* Fixed width */
     padding: 20px; /* Optional padding for the form */
     border: 1px solid #ccc; /* Optional border */
-    border-radius: 8px; /* Optional rounded corners */
+
   }
 
 .input-group{
@@ -39,10 +41,11 @@
     justify-content: center;
     margin-top: 20px;
     width: 100%;
-  }
-  
+}
+
 .login-button{
     width: 100px;
+    border-radius: 100px;
     padding: 10px;
     cursor: pointer;
 }
